+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Changed
+
+- Better error handling, including support for when GitHub rate-limiting errors
+  are provided with non-error (200) codes.
+
 ## 1.7.1 - 2021-10-21
 
 ### Added